refactor: move calendar class to ui module
